Round2 B
Given the probability that each one votes yes, and no one abstains.
Choose K(which is even) people, maximum the proability of tie.

Q: How to choose this K people?
A: No harm to sort, then choose the extrems.
Why? Given K-1 chosen, how to choose the last one?
the last one will be a linear function of p_last, which means p_last needs to be the min or max one from the set of unchosen. So we can only choose from the two ends.